api/requirements.txt
# PlayaTewsIdentityMasker API Requirements
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP client for testing
requests==2.31.0

# Image processing (for file validation)
Pillow==10.1.0

# Optional: Database support (for production)
# sqlalchemy==2.0.23
# alembic==1.13.1

# Optional: Redis for caching (for production)
# redis==5.0.1

# Optional: Monitoring and metrics
# prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
python-json-logger==2.0.7

# CORS support
python-cors==1.7.0

# Environment variable management
python-dotenv==1.0.0 